+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix a freeze after clicking button to add a toolchain. [#37](https://github.com/vlang/intellij-v/issues/37)
 - Allow newlines in between operators in expressions. [#34](https://github.com/vlang/intellij-v/issues/34)
 - Introduce links in console when program panic's and shows backtrace. [#12](https://github.com/vlang/intellij-v/issues/12)
+- Basic support for static methods. [#24](https://github.com/vlang/intellij-v/issues/24)
+  - Please file issues for missing features.
 
 ## [2025.2.1] - 2025-08-09